pyproject.toml
[tool.poetry]
name = "plapobot"
version = "0.9.0"
description = "PLAnning POker Bot - just configure it, add to Slack then enjoy!"
authors = ["Sergey Pechenko a.k.a. tnt4brain"]
license = "MIT License"

[tool.poetry.dependencies]
python = "^3.8"
Flask = "^2.2.2"
Jinja2 = "^3.1.2"
MarkupSafe = "^2.1.1"
PyYAML = "^6.0.0"
slack-bolt = "^1.14.3"
slack-sdk = "^3.18.1"
slackclient = "^2.9.4"
slackeventsapi = "^3.0.1"
Werkzeug = "^2.2.2"
aiosignal = "^1.2.0"
chardet = "^5.0.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
